pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "google-cloud-vision>=3.10.2",
    "httpx>=0.28.1",
    "openai>=1.88.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]
